refactoring, fixing, testing
